Move BackgroundRuns into agent/mason/background.py
The background-run store is a non-durable stand-in slated for a durable backend, so it belongs with
the other mason plumbing. server/app.py imports BackgroundRuns from mason (this diverges its
server/app.py from agent-openai-scratch's).
